C++ API: (CLEC) - move functions L_cc2(), L_cc_stream(), L_cc() and                   L_solve() as CLEC constructors                 - rename function L_cc1() as generate_lec_expression()                 - rename source file l_cc1.c as l_compile.c                 - delete source files l_cc2.c and l_err.c
part of issue #1029

+++ api/lec/l_exec_fns.h
@@ -228,7 +228,7 @@ public:
         pos_buffer++;
     }
 
-    int get_length() const override 
+    short get_length() const override 
     {
         return 2;   // 1 byte for type and 1 byte for nb_args
     }
