+++ api/lec/l_exec_ops.h
@@ -106,7 +106,7 @@ public:
         return is_same_type(other);
     }
 
-    int get_length() const override 
+    short get_length() const override 
     {
         return 1;
     }
